backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.3.1",
    "django>=5.2.3,<6.0",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-phonenumber-field[phonenumberslite]>=8.4.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "djoser>=2.3.1",
    "factory-boy>=3.3.3",
    "flake8>=7.3.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
]

[tool.black]
line-length = 88
target-version = ["py313"]
